move tests into src
under __tests__ directories

also use `craco test` so that we're using the same babel configuration in tests as we do in development

also suppress the "not wrapped in act" warning
