--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  coverageReporters: ['json', 'lcov', 'text-summary'],
-  collectCoverageFrom: ['src/**/*.js'],
-  roots: ['src', 'test']
-};
