--- test/index.spec.js
@@ -1,3 +1,0 @@
-test('self-test', () => {
-  expect(11).toEqual(11);
-});
